Temporary fix for cursor position on reload bug.
Reload character widths calculation after load so cursor is repainted, which
fixes the bug caused by using antialiased font in this theme. Since
observeTextEditor doesn't wait till DOM ready, instead attached to the 1st
movement of the cursor and immediately dispose. This is a band-aid solution
so eventually implement a better solution.