Significant improvements to Semantic Highlighting.
* DRY up semantic highlighting methods.
* Now accurately can determine what word is under the cursor, and
make other occurences of that word which are visible glow.

Initially, semantic highlighting was just a secondary feature I added
in order to fill the gap where so many things were not scoped yet.
Even now, semantic highlighting only applies to text that whatever
grammar is loaded has left unscoped. Because of this, and because
scoping from grammars is much more thorough now, people might
consider this 'broken'.