Fixes #17 & more - Independently manage editors & events.
* Made a lot of changes to the way the view is handled. First
of all, rather than using a single observer for each config setting which
has to perform initialization as well as react to config changes in a
single callback, a single 'observeTextEditors' event is listened upon,
and in its callback we check and apply decoration based on the config,
so the only responsibility is to initalize within that callback.

This allows us to use 'onDidChange' rather than observers for callbacks
which handle individual config values, which allows us to decouple a lot
of stuff and overall provides a more stable, cleaner approach.

* Adds support for scoped wrapping. Whereas before, all
decorations had to be either enabled or disabled for all text editors,
the new approach treats each editor individually.

* Fixes the content padding not going away when turining off wrapping.

* Fixed cursor going off-screen when wraps are off but padding is on.

* Improved performance, but maybe will have issues with semantic
highlighting.

# 50-character subject line
#
# 72-character wrapped longer description. This should answer:
#
# * Why was this change necessary?
# * How does it address the problem?
# * Are there any side effects?
#
# Include a link to the ticket, if any.